[Widgets] Moving widget out of Widgets.[h/cc] and putting it in a 'common' file
